move server directory
