--- server/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "bike-kollective-server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "tsc && nodemon server.ts",
-    "start": "node ./dist/index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.18.1",
-    "firebase": "^9.8.4"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.13",
-    "nodemon": "^2.0.19",
-    "ts-node": "^10.8.2",
-    "typescript": "^4.7.4"
-  }
-}
